Fixed the API code to where it is now functional
